+++ examples/simple.js
@@ -1,0 +1,7 @@
+"use strict";
+
+var geo = require('../lib/geosearch');
+var ip = process.argv[2] || '8.8.8.8';
+geo.lookup(ip, function (err, res) {
+    console.log(err, res);
+});